Move report-related errors code to errors/test-run and the runtime errors code to errors/runtime (#347 part 1)
