+++ src/cli/argument-parser.js
@@ -10,8 +10,8 @@ import OS from 'os-family';
 import { getInstallations as getBrowserInstallations } from 'testcafe-browser-natives';
 import { readSync as read } from 'read-file-relative';
 import promisify from '../utils/promisify';
-import { GeneralError } from '../errors';
-import MESSAGE from '../errors/message';
+import { GeneralError } from '../errors/runtime';
+import MESSAGE from '../errors/runtime/message';
 import getViewPortWidth from '../utils/get-viewport-width';
 import wordWrap from '../utils/word-wrap';
