+++ src/errors/test-run/assertion-diffs/truncate-diff.js
